Updates to comments  and test name changes.
