+++ Samples/ResourceManagement/DocumentDB/CreateDocumentDBWithEventualConsistency/CreateDocumentDBWithEventualConsistency.xproj
@@ -4,14 +4,16 @@
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">14.0</VisualStudioVersion>
     <VSToolsPath Condition="'$(VSToolsPath)' == ''">$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)</VSToolsPath>
   </PropertyGroup>
+
   <Import Project="$(VSToolsPath)\DotNet\Microsoft.DotNet.Props" Condition="'$(VSToolsPath)' != ''" />
   <PropertyGroup Label="Globals">
-    <ProjectGuid>32a300d2-7dda-4b84-9903-522fea0d075c</ProjectGuid>
-    <RootNamespace>HADocumentDB</RootNamespace>
+    <ProjectGuid>AD18F398-E2F1-48BE-94FA-71F31393BA12</ProjectGuid>
+    <RootNamespace>CreateDocumentDBWithEventualConsistency</RootNamespace>
     <BaseIntermediateOutputPath Condition="'$(BaseIntermediateOutputPath)'=='' ">.\obj</BaseIntermediateOutputPath>
     <OutputPath Condition="'$(OutputPath)'=='' ">.\bin\</OutputPath>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
   </PropertyGroup>
+
   <PropertyGroup>
     <SchemaVersion>2.0</SchemaVersion>
   </PropertyGroup>
