+++ Samples/ResourceManagement/DocumentDB/CreateDocumentDBWithEventualConsistency/Properties/AssemblyInfo.cs
