+++ Samples/ResourceManagement/DocumentDB/CreateDocumentDBWithEventualConsistency/project.json
