+++ Samples/ResourceManagement/DocumentDB/CreateDocumentDBWithIPRange/Properties/AssemblyInfo.cs
