+++ Samples/ResourceManagement/DocumentDB/CreateDocumentDBWithKindMongoDB/Properties/AssemblyInfo.cs
