+++ Samples/ResourceManagement/DocumentDB/CreateDocumentDBWithKindMongoDB/project.json
