+++ Samples/ResourceManagement/DocumentDB/ManageHADocumentDB/Properties/AssemblyInfo.cs
