+++ Samples/ResourceManagement/DocumentDB/ManageHADocumentDB/project.json
